associaton added ifelse mi file exists

+++ association/association.py
@@ -101,11 +101,14 @@ if __name__ == '__main__':
         pp = diamond_to_pp(diamond,genomes)
         pickle.dump(pp, open(fileout_pp,'wb'))
         print('saving ' + fileout_pp)
+    else:
+        pp = pd.read_pickle(open(fileout_pp,'rb'))
+        print(fileout_pp +' exists, skipping')
 
+
+    if not os.path.exists(fileout_mi):
         mi = pp_metadata_association(pp,metadata[metadata_colname])
         pickle.dump(mi, open(fileout_mi,'wb'))
         print('saving ' + fileout_mi)
 
         # diamond = read_diamond_output(f)
-    else:
-        print(fileout_pp +' exists, skipping')
